+++ changelog.md
@@ -6,6 +6,7 @@
 - Added Matrix4.isZero()
 - Added Matrix3.isIdentity()
 - Added Matrix3.isZero()
+- Added Vector2.angleTo and Vector2.angleToSigned
 
 ## v 2.0.3 - May 2016